+++ daskms/tests/test_ms_read_and_update.py
@@ -14,6 +14,7 @@ try:
 except ImportError:
     from dask.utils import key_split
 
+from daskms.reads import PARTITION_KEY
 from daskms.utils import (group_cols_str, index_cols_str,
                           select_cols_str, assert_liveness,
                           table_path_split)
@@ -289,7 +290,8 @@ def test_column_promotion(ms):
 
     for ds in xds:
         assert "DATA" in ds.data_vars
-        assert list(ds.attrs.keys()) == ["SCAN_NUMBER"]
+        assert "SCAN_NUMBER" in ds.attrs
+        assert ds.attrs[PARTITION_KEY] == (("SCAN_NUMBER", "int32"),)
 
 
 def test_read_array_names(ms):