+++ HISTORY.rst
@@ -4,11 +4,13 @@ History
 
 0.2.7 (YYYY-MM-DD)
 ------------------
-* Experimental zarr support (:pr:`129`)
+* Correct dataset persistence issues (:pr:`140`)
+* Experimental arrow support (:pr:`130`, :pr:`132`, :pr:`133`, :pr:`135`, :pr:`136`, :pr:`138`)
+* Experimental zarr support (:pr:`129`, :pr:`133`, :pr:`139`)
 * Test data fix (:pr:`128`)
 * Fix array inlining for writes (:pr:`126`)
 * Allow Multi-Layer Inlining (:pr:`125`)
-* Support DATA Column Expressions (:pr:`124`)
+* Support DATA Column Expressions (:pr:`124`, :pr:`134`)
 
 
 0.2.6 (2020-10-20)